fix(product-detail): merge to Detail page
