+++ src/pages/Detail/components/DescriptionTabs/DescriptionTabs.scss
