+++ src/pages/Detail/components/DescriptionTabs/DescriptionTabs.vue
