+++ src/pages/Detail/components/Product/Product.scss
