+++ src/pages/Detail/components/Product/Product.vue
