+++ src/pages/Detail/components/RelatedProducts/RelatedProducts.vue
