+++ src/pages/Detail/components/Sidebar/Sidebar.scss
