+++ src/pages/Detail/components/Sidebar/Sidebar.vue
